Tests/LibWeb/Text/expected/scrollable-viewport-with-position-fixed.txt

{"x":8,"y":8,"width":784,"height":18,"top":8,"right":792,"bottom":26,"left":8}
